:recycle: refactor: clean up event component import/exports
