+++ frontend/src/components/index.ts
@@ -1,4 +1,6 @@
 export { Breadcrumbs } from "./Breadcrumbs";
 export { Markdown } from "./Markdown";
 export { Logo } from "./Logo";
-export { default as Link, type LinkProps, NextLinkComposed } from "./Link";
+export { default as Link, type LinkProps } from "./Link";
+export { Title } from "./Title";
+export { PermissionRequired, LoginRequired } from "./Auth";
