+++ frontend/src/components/pages/events/filterMenu/DateTimeFilter.tsx
@@ -1,7 +1,7 @@
 import { Grid, TextField } from "@mui/material";
 import React from "react";
 
-import { FilterQuery } from "@/components/pages/events/AllEvents";
+import { FilterQuery } from "../AllEvents";
 
 type Props = {
   /** The currently applied filters */
