+++ frontend/src/components/pages/events/filterMenu/index.tsx
@@ -2,7 +2,7 @@ import { Refresh, StarOutlineRounded, StarRounded } from "@mui/icons-material";
 import { Card, CardContent, Divider, Grid, IconButton, Tooltip, Typography } from "@mui/material";
 import React from "react";
 
-import { FilterQuery } from "@/components/pages/events/AllEvents";
+import { FilterQuery } from "../AllEvents";
 
 import { CategoryFilter } from "./CategoryFilter";
 import { DateTimeFilter } from "./DateTimeFilter";
